requirements.txt
yt-dlp>=2024.0.0
mutagen>=1.47.0
requests>=2.31.0
tqdm>=4.66.0
spotipy>=2.23.0
python-dotenv>=1.0.0